# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TukPatrol API",
    "version": "1.0.0",
    "description": "Real-Time Three-Wheeler Tracking and Movement Logging System for Sri Lanka Law Enforcement"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Province": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "Western"
          },
          "code": {
            "type": "string",
            "example": "WP"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "District": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "Colombo"
          },
          "code": {
            "type": "string",
            "example": "COL"
          },
          "province": {
            "$ref": "#/components/schemas/Province"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PoliceStation": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "Colombo Fort Police Station"
          },
          "code": {
            "type": "string",
            "example": "ST001"
          },
          "district": {
            "$ref": "#/components/schemas/District"
          },
          "province": {
            "$ref": "#/components/schemas/Province"
          },
          "address": {
            "type": "string"
          },
          "contactNumber": {
            "type": "string"
          }
        }
      },
...
      "Vehicle": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "registrationNumber": {
            "type": "string",
            "example": "WP-0001"
          },
          "deviceId": {
            "type": "string",
            "example": "DEV-0001"
          },
          "driverName": {
            "type": "string",
            "example": "Kamal Perera"
          },
          "driverNIC": {
            "type": "string",
            "example": "198512340001V"
          },
          "driverContact": {
            "type": "string",
            "example": "0771234567"
          },
          "province": {
            "$ref": "#/components/schemas/Province"
          },
          "district": {
            "$ref": "#/components/schemas/District"
          },
          "station": {
            "$ref": "#/components/schemas/PoliceStation"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "LocationPing": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "vehicle": {
            "type": "string"
          },
          "latitude": {
            "type": "number",
            "example": 6.9271
          },
          "longitude": {
            "type": "number",
            "example": 79.8612
          },
          "speed": {
            "type": "number",
            "example": 35.5
          },
          "heading": {
            "type": "number",
            "example": 180
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "code",
          "message",
          "description",
          "moreInfo"
        ],
        "properties": {
          "code": {
            "type": "string",
            "example": "401"
          },
          "message": {
            "type": "string",
            "example": "Unauthorized"
          },
          "description": {
            "type": "string",
            "example": "Not authorized, no token"
          },
          "moreInfo": {
            "type": "string",
            "example": ""
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/vehicles": {
      "get": {
        "summary": "Get all vehicles",
        "tags": [
          "Vehicles"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "province",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "district",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "station",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "isActive",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of vehicles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "page": {
                      "type": "integer"
                    },
                    "total": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Vehicle"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Register a new vehicle",
        "tags": [
          "Vehicles"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "registrationNumber",
                  "deviceId",
                  "driverName",
                  "driverNIC",
                  "province",
                  "district",
                  "station"
                ],
                "properties": {
                  "registrationNumber": {
                    "type": "string",
                    "example": "WP-0001"
                  },
                  "deviceId": {
                    "type": "string",
                    "example": "DEV-0001"
                  },
                  "driverName": {
                    "type": "string",
                    "example": "Kamal Perera"
                  },
                  "driverNIC": {
                    "type": "string",
                    "example": "198500100001V"
                  },
                  "driverContact": {
                    "type": "string",
                    "example": "0771234567"
                  },
                  "province": {
                    "type": "string",
                    "example": "60d0fe4f5311236168a109ca"
                  },
                  "district": {
                    "type": "string",
                    "example": "60d0fe4f5311236168a109cb"
                  },
                  "station": {
                    "type": "string",
                    "example": "60d0fe4f5311236168a109cc"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Vehicle created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vehicle"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vehicles/{id}": {
      "get": {
        "summary": "Get a standard vehicle",
        "tags": [
          "Vehicles"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vehicle details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vehicle"
                }
              }
            }
          },
          "404": {
            "description": "Vehicle not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a vehicle",
        "tags": [
          "Vehicles"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Vehicle updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vehicle"
                }
              }
            }
          },
          "404": {
            "description": "Vehicle not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a vehicle",
        "tags": [
          "Vehicles"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Vehicle deleted"
          },
          "404": {
            "description": "Vehicle not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vehicles/{id}/deactivate": {
      "put": {
        "summary": "Deactivate a vehicle",
        "tags": [
          "Vehicles"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vehicle deactivated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Vehicle deactivated"
                    },
                    "vehicle": {
                      "$ref": "#/components/schemas/Vehicle"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Vehicle not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/vehicles/{id}/location": {
      "get": {
        "summary": "Get last known location of a vehicle",
        "tags": [
          "Vehicles"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Last known location",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "vehicle": {
                      "$ref": "#/components/schemas/Vehicle"
                    },
                    "lastLocation": {
                      "$ref": "#/components/schemas/LocationPing"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No location data found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vehicles/{id}/history": {
      "get": {
        "summary": "Get location history of a vehicle",
        "tags": [
          "Vehicles"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "to",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Location history",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "page": {
                      "type": "integer"
                    },
                    "total": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/LocationPing"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Vehicle not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
...
}